feat(editor): support vscode terminal
When using VSCode, I wanted to be able to use `git` commands from the
embedded terminal. To do so, we detect if the terminal is running in
VSCode using the $TERM_PROGRAM variable. That allows me to use thing
like `git commit` inside the VSCode embedded terminal and it will use
that same VSCode instance as the $GIT_EDITOR